+++ src/BlockHorizons/BlockSniper/ui/windows/PresetCreationWindow.php
@@ -7,54 +7,40 @@ namespace BlockHorizons\BlockSniper\ui\windows;
 use BlockHorizons\BlockSniper\brush\BrushProperties;
 use BlockHorizons\BlockSniper\brush\registration\ShapeRegistration;
 use BlockHorizons\BlockSniper\brush\registration\TypeRegistration;
-use BlockHorizons\BlockSniper\brush\shapes\SphereShape;
-use BlockHorizons\BlockSniper\brush\types\FillType;
 use BlockHorizons\BlockSniper\data\Translation;
 use BlockHorizons\BlockSniper\Loader;
 use BlockHorizons\BlockSniper\presets\Preset;
+use BlockHorizons\BlockSniper\sessions\SessionManager;
 use pocketmine\Player;
 
 class PresetCreationWindow extends CustomWindow{
 
 	/** @var Preset */
-	private $preset;
+	public $preset;
 
 	public function __construct(Loader $loader, Player $requester){
-		parent::__construct($this->t(Translation::UI_PRESET_MENU_TITLE));
+		parent::__construct($this->t(Translation::UI_PRESET_CREATION_TITLE));
 		$this->preset = new Preset("", new BrushProperties());
+		$b = SessionManager::getPlayerSession($requester)->getBrush();
 
 		$this->addInput($this->t(Translation::UI_PRESET_CREATION_NAME), "New Preset", $this->t(Translation::UI_PRESET_CREATION_NAME), function(Player $player, string $value){
 			$this->preset->name = $value;
 		});
-		$this->addSlider($this->t(Translation::UI_PRESET_CREATION_SIZE), 0, $loader->config->maxSize, 1, 10, function(Player $player, float $value){
-			$this->preset->properties->size = (int) $value;
+		$this->addDropdown($this->t(Translation::UI_BRUSH_MENU_MODE_DESCRIPTION), $this->processModes(), $b->mode, function(Player $player, int $value) {
+			$this->preset->properties->mode = $value;
 		});
-		$this->addDropdown($this->t(Translation::UI_PRESET_CREATION_SHAPE), $this->processShapes($requester), SphereShape::ID, function(Player $player, int $value){
+		$this->addDropdown($this->t(Translation::UI_PRESET_CREATION_SHAPE), $this->processShapes($requester), $b->getShape()::ID, function(Player $player, int $value){
 			$this->preset->properties->shape = ShapeRegistration::getShapeById($value);
 		});
-		$this->addDropdown($this->t(Translation::UI_PRESET_CREATION_TYPE), $this->processTypes($requester), FillType::ID, function(Player $player, int $value){
+		$this->addDropdown($this->t(Translation::UI_PRESET_CREATION_TYPE), $this->processTypes($requester), $b->getType()::ID, function(Player $player, int $value) use($loader) {
 			$this->preset->properties->type = TypeRegistration::getTypeById($value);
-		});
-		$this->addToggle($this->t(Translation::UI_PRESET_CREATION_HOLLOW), false, function(Player $player, bool $value){
-			$this->preset->properties->hollow = $value;
-		});
-		$this->addToggle($this->t(Translation::UI_PRESET_CREATION_DECREMENT), false, function(Player $player, bool $value){
-			$this->preset->properties->decrementing = $value;
-		});
-		$this->addSlider($this->t(Translation::UI_PRESET_CREATION_HEIGHT), 0, $loader->config->maxSize, 1, 10, function(Player $player, float $value){
-			$this->preset->properties->height = (int) $value;
-		});
-		$this->addInput($this->t(Translation::UI_PRESET_CREATION_BLOCKS), "stone", "stone,stone_brick:1,2", function(Player $player, string $value){
-			$this->preset->properties->blocks = $value;
-		});
-		$this->addInput($this->t(Translation::UI_PRESET_CREATION_OBSOLETE), "air", "stone,stone_brick:1,2", function(Player $player, string $value){
-			$this->preset->properties->obsolete = $value;
-		});
-		$this->addInput($this->t(Translation::UI_PRESET_CREATION_BIOME), "plains", "plains", function(Player $player, string $value) use($loader){
-			$this->preset->properties->biomeId = $this->preset->properties->parseBiomeId($value);
-			// Last element, so we finish the preset here and add it.
-			if(!$loader->getPresetManager()->isPreset($this->preset->name)){
-				$loader->getPresetManager()->addPreset($this->preset);
+			$loader->getPresetManager()->addPreset($this->preset);
+
+			// Note to future readers: This response form is explicitly set after all preset properties have been
+			// changed, as we want to send the requester back to the main menu if they press the cross.
+			$form = new PresetPropertiesWindow($loader, $player, $this->preset);
+			if($form->elementCount() !== 0){
+				$this->setResponseForm($form);
 			}
 		});
 		$this->setResponseForm(new PresetMenuWindow($loader, $requester));