+++ CHANGELOG.md
@@ -3,11 +3,14 @@
 ### New Features
 - Added a new UI that opens after setting the brush shape, type and mode. This new UI contains brush properties
   specifically available for the combination of shape, type and mode set.
+- Presets can now set tree properties in addition to other properties.
   
 ### Behavioural Changes
 - The Brush UI no longer contains all available brush properties.
 - The main menu no longer contains a Tree button.
 - The tree properties now show when the new brush properties UI is opened.
+- The preset creation window no longer holds all available properties. Instead, like the brush window, it now only shows
+  brush properties that apply.
 
 ### Bug Fixes
 - Fixed the Clean type doing the opposite of what it is supposed to do.